take5/requirements.txt
Django==4.1.2
djangorestframework==3.14.0
flake8==5.0.4
python-decouple==3.6
